Ensure DOM is cleaned up when `BlockTracker#destroy` is called.
This is currently called by Ember for the top level view, and is
responsible for cleaning up any remaining DOM. The previous
implementation would ensure that any created views were destroyed (and
their various hooks were called), but it did not actually remove the DOM
at all. This made the tests in Ember leak DOM (and ultimately fail in
odd ways because of duplicate DOM elements).

It is important to note that `BlockTracker.prototype.destroy` and
`BlockTracker.prototype.reset` are _extremely_ similar as of this
change. The only difference is that `.reset` queues destructor using the
`Environment.prototype.didDestroy` hook, and `.destroy` calls
`.destroy` on the items immediately.